optimizer/converter: rename to optimizer/layerconverter
